backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
bcrypt==4.2.0
python-multipart==0.0.9
pydantic>=2.0,<3.0
openpyxl==3.1.5
httpx==0.27.0
python-dotenv==1.0.1
python3-saml==1.16.0
psycopg2-binary>=2.9.0
PyJWT==2.10.1
